Extract StringValue into string_value.dart
