+++ lib/src/sdk/boards/board_id.dart
@@ -1,4 +1,4 @@
-import '../client.dart' show StringValue;
+import '../string_value.dart' show StringValue;
 
 class BoardId extends StringValue {
   BoardId(String id) : super(id);
