+++ lib/src/sdk/cards/attachment_id.dart
@@ -1,4 +1,4 @@
-import '../client.dart' show StringValue;
+import '../string_value.dart' show StringValue;
 
 class AttachmentId extends StringValue {
   AttachmentId(super.id);
