+++ lib/src/sdk/cards/card_id.dart
@@ -1,4 +1,4 @@
-import '../client.dart' show StringValue;
+import '../string_value.dart' show StringValue;
 
 class CardId extends StringValue {
   CardId(super.id);
