+++ lib/src/sdk/file_name.dart
@@ -1,4 +1,4 @@
-import 'client.dart';
+import 'string_value.dart' show StringValue;
 
 class FileName extends StringValue {
   FileName(super.fileName);
