+++ lib/src/sdk/lists/list_id.dart
@@ -1,4 +1,4 @@
-import '../client.dart' show StringValue;
+import '../string_value.dart' show StringValue;
 
 class ListId extends StringValue {
   ListId(id) : super(id);
