+++ lib/src/sdk/members/member_id.dart
@@ -1,4 +1,4 @@
-import '../client.dart' show StringValue;
+import '../string_value.dart' show StringValue;
 
 class MemberId extends StringValue {
   MemberId(super.id);
